refactor(delegate): restructure TypeScript tools into barrel subdirectory
